fix(search-and-filter) custom event for panel toggle, simplify menu hook
